+++ CHANGELOG.md
@@ -16,8 +16,10 @@
 ### Fixes
 
 * Raise error (not show strange error) while trying to create a lot of server
+* Fix setting default value for `spec_browser`
 
 ### Changes
 
 * Require ruby version >= 2.4
 * Remove usage of `cleanup_test_path` while adding tests to queue
+* `WrataApi#tests_in_queue` return all test data, not just name